admin restrictions i views.py

+++ Captain_Console/products/views.py
@@ -9,15 +9,11 @@ from orders.models import Order, OrderProduct
 from products.forms.product_form import ProductForm
 from products.forms.image_form import ImageForm
 from products.forms.review_form import ReviewForm
-
 from products.models import Product, ProductImage, ProductHistory, Review
-
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404
 from django.http import JsonResponse
-
 from users.models import Customer
-
 
 def frontpage(request):
     if 'search_filter' in request.GET:
@@ -31,6 +27,7 @@ def frontpage(request):
             'image': ProductImage.objects.filter(product=x.id).first().image
         } for x in Product.objects.filter(name__icontains=search_filter)]
         return JsonResponse({'data': products})
+
     profile = None
     if request.user.is_authenticated:
         profile = Customer.objects.get(user=request.user)
@@ -97,39 +94,45 @@ def get_product_by_id(request, id, consolename=None, name=None):
     return render(request, 'products/product_details.html', product)
 
 def create_product(request):
-    if request.method == "POST":
-        form1 = ProductForm(data=request.POST)
-        if form1.is_valid():
-            the_cons = Console.objects.get(pk=form1.instance.console_type.id)
-            form1.instance.manufacturer = Manufacturer.objects.get(pk=the_cons.manufacturer.id)
-            form2 = ImageForm(data=request.POST)
-            form2.instance.product = form1.instance
-            form2.save()
-            return redirect('products')
-    return render(request, 'products/create_product.html', {
-        'form1': ProductForm(),
-        'form2': ImageForm()
-    })
+
+    if request.user.is_superuser:
+        if request.method == "POST":
+            form1 = ProductForm(data=request.POST)
+            if form1.is_valid():
+                the_cons = Console.objects.get(pk=form1.instance.console_type.id)
+                form1.instance.manufacturer = Manufacturer.objects.get(pk=the_cons.manufacturer.id)
+                form2 = ImageForm(data=request.POST)
+                form2.instance.product = form1.instance
+                form2.save()
+                return redirect('products')
+        return render(request, 'products/create_product.html', {
+            'form1': ProductForm(),
+            'form2': ImageForm()
+        })
 
 def update_product(request, id):
-    the_product = Product.objects.filter(pk=id).first()
-    if request.method == "POST":
-        form = ProductForm(data=request.POST,instance=the_product)
-        if form.is_valid():
-            the_cons = Console.objects.get(pk=form.instance.console_type.id)
-            form.instance.manufacturer = Manufacturer.objects.get(pk=the_cons.manufacturer.id)
-            form.save()
-            return redirect('products')
-    return render(request, 'products/update_product.html', {
-        'form': ProductForm(instance=the_product)
-    })
+
+    if request.user.is_superuser:
+        the_product = Product.objects.filter(pk=id).first()
+        if request.method == "POST":
+            form = ProductForm(data=request.POST,instance=the_product)
+            if form.is_valid():
+                the_cons = Console.objects.get(pk=form.instance.console_type.id)
+                form.instance.manufacturer = Manufacturer.objects.get(pk=the_cons.manufacturer.id)
+                form.save()
+                return redirect('products')
+        return render(request, 'products/update_product.html', {
+            'form': ProductForm(instance=the_product)
+        })
 
 def delete_product(request, id):
-    the_product = Product.objects.filter(pk=id).first()
-    the_product.delete()
-    return render(request, 'products/delete_product.html', {
-        'form': ProductForm(instance=the_product)
-    })
+
+    if request.user.is_superuser:
+        the_product = Product.objects.filter(pk=id).first()
+        the_product.delete()
+        return render(request, 'products/delete_product.html', {
+            'form': ProductForm(instance=the_product)
+        })
 
 @login_required()
 def review_product(request, id):
